refactored participant app
